Move configurable fields models and graphql in their own python modules
