+++ assembl/models/__init__.py
@@ -122,7 +122,9 @@ from .auth import (  # noqa: E402, F401
     UserLanguagePreference,
     UserRole,
     UserTemplate,
-    Username,
+    Username
+)
+from .configurable_fields import (  # noqa: E402, F401
     TextField,
     ProfileTextField
 )
